Rename BaseDeviceValue to BaseValue and use double (instead of float) everywhere in units, values,...

+++ IHA/src/cz/vutbr/fit/iha/activity/SensorDetailFragment.java
@@ -464,7 +464,7 @@ public class SensorDetailFragment extends SherlockFragment {
 					return formatter.print((long) value);
 				}
 
-				return String.format("%s %s", fmt.getStringValue(mDevice.getValue(), (float)value), unit);
+				return String.format("%s %s", fmt.getStringValue(mDevice.getValue(), value), unit);
 			}
 		});
