+++ IHA/src/cz/vutbr/fit/iha/adapter/device/values/OnOffValue.java
@@ -30,8 +30,8 @@ public class OnOffValue extends BaseEnumValue {
 	}
 
 	@Override
-	public float getFloatValue() {
-		return Float.NaN;
+	public double getDoubleValue() {
+		return Double.NaN;
 	}
 
 	private boolean isActive() {
